Fixed cloning URL
It was pointing to an old repo (fork?). Just changed it to the current repository, with so many copies of this going around this was the one thing keeping me from having this work in my machine. An easy thing to overlook, especially by a user who's just copying and pasting.